First commit with full destructive delete

+++ routes/application.php
@@ -4,17 +4,63 @@
 $app->get(
 	'/',
 	function() use ($app,$appConf) {
-        $openStack = $app->openStack;
+    $openStack = $app->openStack;
 
 		$compute = $openStack->computeV2();
-		$servers = $compute->listServers(true);
+		$ocpNodes = $compute->listServers(true, ['flavorId'=>'4d3d73a3-4575-49a1-93d3-2a20c565aded']);
+		$ocpControlNodes = $compute->listServers(true, ['flavorId'=>'50503afe-7a3e-4768-9a35-3f097264d6ee']);
 
-		foreach ($servers as $server) {
-			print_r($server);
+		$ocpControlStatus = array();
+		$ocpNodeStatus = array();
+
+		foreach ($ocpControlNodes as $node) {
+			$nodeStatus = array("type"=>"Master/Infra","id"=>$node->id,"name"=>$node->name,"status"=>$node->status);
+			array_push($ocpControlStatus,$nodeStatus);
+		}
+
+		foreach ($ocpNodes as $node) {
+			$nodeStatus = array("type"=>"Node","id"=>$node->id,"name"=>$node->name,"status"=>$node->status);
+			array_push($ocpNodeStatus,$nodeStatus);
 		}
 
 		$app->render(
-			'index.html'
+			'index.html',
+			array(
+				'ocpControlStatus'=>$ocpControlStatus,
+				'ocpNodeStatus'=>$ocpNodeStatus
+			)
+		);
+	}
+);
+
+$app->get(
+	'/ocp-cluster-status',
+	function() use ($app,$appConf) {
+    $openStack = $app->openStack;
+
+		$compute = $openStack->computeV2();
+		$ocpNodes = $compute->listServers(true, ['flavorId'=>'4d3d73a3-4575-49a1-93d3-2a20c565aded']);
+		$ocpControlNodes = $compute->listServers(true, ['flavorId'=>'50503afe-7a3e-4768-9a35-3f097264d6ee']);
+
+		$ocpControlStatus = array();
+		$ocpNodeStatus = array();
+
+		foreach ($ocpControlNodes as $node) {
+			$nodeStatus = array("type"=>"Master/Infra","id"=>$node->id,"name"=>$node->name,"status"=>$node->status);
+			array_push($ocpControlStatus,$nodeStatus);
+		}
+
+		foreach ($ocpNodes as $node) {
+			$nodeStatus = array("type"=>"Node","id"=>$node->id,"name"=>$node->name,"status"=>$node->status);
+			array_push($ocpNodeStatus,$nodeStatus);
+		}
+
+		$app->render(
+			'ocpClusterStatus.html',
+			array(
+				'ocpControlStatus'=>$ocpControlStatus,
+				'ocpNodeStatus'=>$ocpNodeStatus
+			)
 		);
 	}
 );
@@ -29,13 +75,38 @@ $app->get(
 		$servers = $compute->listServers(true, ['flavorId'=>'4d3d73a3-4575-49a1-93d3-2a20c565aded']);
 
 		foreach ($servers as $server) {
-        	array_push($serverArr, $server);
-        }
-
-        $response = $app->response();
-    	$response['Content-Type'] = 'application/json';
-    	$response->body(json_encode($serverArr));
+    	array_push($serverArr, $server);
     }
+
+    $response = $app->response();
+  	$response['Content-Type'] = 'application/json';
+  	$response->body(json_encode($serverArr));
+  }
+);
+
+$app->get(
+	'/random-ocp-node-id',
+	function() use ($app, $appConf) {
+		$serverArr = array();
+		$openStack = $app->openStack;
+
+		$compute = $openStack->computeV2();
+		$servers = $compute->listServers(true, ['flavorId'=>'4d3d73a3-4575-49a1-93d3-2a20c565aded']);
+
+		$nodeIds = array();
+		foreach ($servers as $server) {
+			array_push($nodeIds, $server->id);
+    }
+
+		$result = false;
+		if (count($nodeIds) > 1) {
+			$result = $nodeIds[array_rand($nodeIds)];
+		}
+
+		$response = $app->response();
+  	$response['Content-Type'] = 'application/json';
+  	$response->body(json_encode($result));
+  }
 );
 
 $app->get(
@@ -68,13 +139,15 @@ $app->get(
 			$servers = $compute->listServers(false, ['flavorId'=>'4d3d73a3-4575-49a1-93d3-2a20c565aded']);
 
 			$isOcpNode = false;
+			$nodeCount = 0;
 			foreach ($servers as $server) {
     		if ($id == $server->id) {
     			$isOcpNode = true;
     		}
+				$nodeCount++;
     	}
 
-    	if ($isOcpNode) {
+    	if ($isOcpNode && $nodeCount > 1) {
     		$server = $compute->getServer(['id' => $id]);
 				$server->delete();
 
@@ -92,7 +165,6 @@ $app->get(
     }
 
     header('Content-Type: application/json');
-
     $response = $app->response();
   	$response['Content-Type'] = 'application/json';
   	$response->body(json_encode($results));
